Omarbetad datahantering. Editorfältet sätter markören i slutet av texten.

+++ src/components/filetools.js
@@ -2,21 +2,37 @@ import React    from 'react';
 import Util     from '../helpers/util';
 import Settings from '../settings.json';
 import Actions  from '../data/treeactions';
+import EditBox  from './editbox';
 
-export default class Fileinfo extends React.Component {
+export default class FileTools extends React.Component {
   constructor(props) {
     super(props);
+    this.state = {
+      file: this.props.file
+    }
   }
+
   handleClick(mode, e) {
     switch (mode) {
       case 'edit':
+        e.preventDefault();
         e.stopPropagation();
-        Actions.setEditmode(this.props.file);
+        Actions.startEditmode(this.props.file);
       break;
       case 'delete':
-        Actions.deleteItem(this.props.file);
+        e.preventDefault();
+        e.stopPropagation();
+        let typeText = 'filen';
+        if (this.props.file.type === 'dir') {
+          typeText = 'katalogen';
+        }
+        if (confirm('Vill du verkligen ta bort ' + typeText + ' ' + this.props.file.name)) {
+          Actions.deleteItem(this.props.file);
+        }
       break;
       case 'done':
+        e.preventDefault();
+        e.stopPropagation();
         Actions.doneEditing(this.props.file);
       break;
       case 'info':
@@ -31,27 +47,37 @@ export default class Fileinfo extends React.Component {
       break;
     }
   }
+
   getStyle(st) {
     if (this.props.file.showInfo) {
       st.color = '#2E9954'
     }
     return st;
   }
+
+  componentWillUpdate() {
+  }
+
   render() {
-    let editTool = <i
-          className = {'fa fa-' + Settings.toolIcon.edit}
-          onClick   = {this.handleClick.bind(this, 'edit')}
-        ></i>
+    let file     = this.state.file;
+    let editTool =
+          <i
+            className = {'fa fa-' + Settings.toolIcon.edit}
+            onClick   = {this.handleClick.bind(this, 'edit')}
+          ></i>
     let deleteTool = <i
           style     = {{paddingLeft:'10px'}}
           className = {'fa fa-' + Settings.toolIcon.delete}
           onClick   = {this.handleClick.bind(this, 'delete')}
         ></i>
-    let doneTool = <i
+    let doneTool = <span>
+      <EditBox file = {file} />
+      <i
           style     = {{paddingLeft:'10px'}}
           className = {'fa fa-' + Settings.toolIcon.done}
           onClick   = {this.handleClick.bind(this, 'done')}
-        ></i>
+      ></i>
+      </span>
     let infoTool = <i
           style     = {this.getStyle({paddingLeft:'10px'})}
           className = {'fa fa-' + Settings.toolIcon.info}
